chore: transfer sections data to data files
