+++ src/components/pages/echo/hero/data.js
@@ -1,0 +1,8 @@
+const DATA = {
+  title: 'Create powerful notification workflows and content using code',
+  description:
+    'Create complex workflows, access local data, and reuse existing content templates with Novu Echo.',
+  link: 'https://novu.mintlify.app/echo/quickstart',
+};
+
+export default DATA;
