+++ src/components/pages/echo/hero/hero.jsx
@@ -4,6 +4,7 @@ import React, { useState, useEffect } from 'react';
 
 import Button from 'components/shared/button';
 
+import DATA from './data';
 import borderGlow from './images/border-glow.svg';
 
 const CODE = 'npx novu-labs@latest echo';
@@ -30,11 +31,10 @@ const Hero = () => {
       <div className="container-xl relative">
         <div className="max-w-[562px] lg:max-w-md md:max-w-none md:px-[68px] sm:px-1">
           <h1 className="text-[52px] leading-denser tracking-snug font-medium lg:text-5xl md:text-4xl sm:text-[30px] sm:text-center">
-            Create powerful notification workflows and&nbsp;content&nbsp;using&nbsp;code
+            {DATA.title}
           </h1>
           <p className="mt-4 text-lg font-book leading-snug text-white/70 md:mt-3 md:text-base sm:text-center">
-            Create complex workflows, access local data, and reuse existing content templates with
-            Novu Echo.
+            {DATA.description}
           </p>
           <div className="flex justify-between mt-12 lg:flex-wrap lg:gap-y-4 lg:mt-[42px] md:justify-start md:gap-x-6 md:mt-9 sm:mt-7">
             <div className="w-[392px] h-[54px] pl-5 pr-2 relative flex items-center justify-between rounded-md bg-black border border-transparent bg-clip-border before:absolute before:-z-20 before:-inset-0.5 before:rounded-md before:bg-[linear-gradient(267.51deg,#FFD447_10.88%,#FF8125_28.58%,#FF036B_62.46%,#DE18BC_109.99%)] lg:h-[46px]">
@@ -59,7 +59,7 @@ const Hero = () => {
             <Button
               className="h-14 -mt-px text-sm min-w-[142px] lg:h-12 lg:mt-0 md:-mt-px"
               theme="gray-outline"
-              to="https://novu.mintlify.app/echo/quickstart"
+              to={DATA.link}
             >
               Learn more
             </Button>
